+++ src/config/mod.rs
@@ -7,6 +7,7 @@ pub mod archive;
 pub mod conf_import;
 pub mod console;
 pub mod defaults;
+pub mod desktop;
 pub mod dosbox_conf;
 pub mod paths;
 pub mod profile;